client: Move worker-related source files to separate directory
